+++ src/BC/Operator/Elastic/Constant.H
@@ -646,7 +646,6 @@ public:
         #endif
 
     }
-
 };
 }
 }
