+++ src/Operator/Elastic.cpp
@@ -829,28 +829,7 @@ Elastic<SYM>::averageDownCoeffsSameAmrLevel(int amrlev)
                     fdata(i, j, k) / 8.0;
 
 #ifdef AMREX_DEBUG
-                if (cdata(I, J, K).contains_nan())
-                {
-                    Util::Warning(INFO,"course lo ",lo);
-                    Util::Warning(INFO, "coarse hi ",hi);
-                    Util::Warning(INFO, "coarse box ",bx);
-                    Util::Warning(INFO);
-                    Util::Warning(INFO,"i=",i,", j=",j);
-                    Util::Warning(INFO);
-                    Util::Warning(INFO,i-1," ",j-1,":\t",fdata(i-1,j-1,k));
-                    Util::Warning(INFO,i-1," ",j,":\t",fdata(i-1,j,k));
-                    Util::Warning(INFO,i-1," ",j+1,":\t",fdata(i-1,j+1,k));
-
-                    Util::Warning(INFO,i," ",j-1,":\t",fdata(i,j-1,k));
-                    Util::Warning(INFO,i," ",j,":\t",fdata(i,j,k));
-                    Util::Warning(INFO,i," ",j+1,":\t",fdata(i,j+1,k));
-
-                    Util::Warning(INFO,i+1," ",j-1,":\t",fdata(i+1,j-1,k));
-                    Util::Warning(INFO,i+1," ",j,":\t",fdata(i+1,j,k));
-                    Util::Warning(INFO,i+1," ",j+1,":\t",fdata(i+1,j+1,k));
-                    
-                    Util::Abort(INFO, "restricted model is nan at crse coordinates (I=", I, ",J=", J, ",K=", k, "), amrlev=", amrlev, " interpolating from mglev", mglev - 1, " to ", mglev);
-                }
+                if (cdata(I, J, K).contains_nan()) Util::Abort(INFO, "restricted model is nan at crse coordinates (I=", I, ",J=", J, ",K=", k, "), amrlev=", amrlev, " interpolating from mglev", mglev - 1, " to ", mglev);
 #endif
             });
         }
