+++ src/Solver/Nonlocal/Newton.H
@@ -60,7 +60,6 @@ private:
         {
             amrex::Box domain(linop->Geom(lev).growPeriodicDomain(1));
             domain.convert(amrex::IntVect::TheNodeVector());
-
             const Set::Scalar* dx = linop->Geom(lev).CellSize();
             Set::Vector DX(linop->Geom(lev).CellSize());
             for (MFIter mfi(*a_model_mf[lev], false); mfi.isValid(); ++mfi)
@@ -99,8 +98,6 @@ private:
                 });
             }
 
-            //a_dw_mf[lev]->FillBoundary(m_elastic->Periodicity(lev));
-            //a_ddw_mf[lev]->FillBoundary(m_elastic->Periodicity(lev));
             Util::RealFillBoundary(*a_dw_mf[lev], m_elastic->Geom(lev));
             Util::RealFillBoundary(*a_ddw_mf[lev], m_elastic->Geom(lev));
         }
@@ -112,7 +109,6 @@ private:
             amrex::Box domain(linop->Geom(lev).growPeriodicDomain(1));
             domain.convert(amrex::IntVect::TheNodeVector());
             const Set::Scalar* dx = linop->Geom(lev).CellSize();
-
             const amrex::Dim3 lo = amrex::lbound(domain), hi = amrex::ubound(domain);
             for (MFIter mfi(*a_model_mf[lev], false); mfi.isValid(); ++mfi)
             {
@@ -156,7 +152,6 @@ private:
         {
             amrex::Box domain(linop->Geom(lev).growPeriodicDomain(2));
             domain.convert(amrex::IntVect::TheNodeVector());
-
             const Set::Scalar* dx = linop->Geom(lev).CellSize();
             Set::Vector DX(linop->Geom(lev).CellSize());
             for (MFIter mfi(*a_model_mf[lev], false); mfi.isValid(); ++mfi)
@@ -205,7 +200,6 @@ private:
         {
             amrex::Box domain(linop->Geom(lev).growPeriodicDomain(2));
             domain.convert(amrex::IntVect::TheNodeVector());
-
             const Set::Scalar* dx = linop->Geom(lev).CellSize();
             const amrex::Dim3 lo = amrex::lbound(domain), hi = amrex::ubound(domain);
             for (MFIter mfi(*a_model_mf[lev], false); mfi.isValid(); ++mfi)
