+++ src/Util/MPI.H
@@ -29,9 +29,9 @@ int Allgather(std::vector<T>& a_data)
     std::vector<T> a_data_all(num);
     // Send information about how many sites on each proc to all procs
     std::vector<int> nsites_procs(nprocs);
-    MPI_Allgather(  &my_num, 1, amrex::ParallelDescriptor::Mpi_typemap<int>::type(),
-                    nsites_procs.data(), 1, amrex::ParallelDescriptor::Mpi_typemap<int>::type(),
-                    MPI_COMM_WORLD);
+    MPI_Allgather(&my_num, 1, amrex::ParallelDescriptor::Mpi_typemap<int>::type(),
+        nsites_procs.data(), 1, amrex::ParallelDescriptor::Mpi_typemap<int>::type(),
+        MPI_COMM_WORLD);
     // Calculate the offset for each
     std::vector<int> nsites_disp(nprocs);
     for (int i = 0; i < nprocs; i++)
@@ -75,7 +75,7 @@ inline void Bcast(  std::string &s,
 
 
 
-inline void AllGather(  const std::string &local,
+inline void Allgather(  const std::string &local,
                         std::vector<std::string> &out)
 {
     using namespace amrex;
