+++ src/Util/Util.cpp
@@ -29,121 +29,6 @@ std::pair<std::string,std::string> file_overwrite;
 bool initialized = false;
 bool finalized = false;
 
-
-std::map<std::string,int>  Compare_map;
-int Compare_cnt = 0;
-
-void Compare(
-    std::string file, std::string func, int line,
-    const amrex::MultiFab &a_mf, std::string desc,
-    amrex::Box domain)
-{
-#ifndef AMREX_DEBUG
-    return;
-#endif
-
-    Compare_cnt++;
-
-    Set::Scalar tolerance = 1E-10;
-
-    std::ostringstream messagestream;
-    std::string filesan = file;
-    Util::String::ReplaceAll(filesan,"/","_");
-    messagestream << filesan << "_" << func << "_" << line << desc;
-    std::string key = messagestream.str();
-
-    int cntr = 0;
-    if (Compare_map.find(key) == Compare_map.end())
-        Compare_map[key] = 0;
-    else
-        Compare_map[key] ++;
-    cntr = Compare_map[key];
-
-    messagestream << "_" << cntr;
-
-    std::string name = messagestream.str();
-    std::vector<std::string> allnames;
-    Util::MPI::AllGather(name,allnames);
-    for (auto & othername : allnames)
-    {
-        if (name != othername)
-            Util::ParallelAbort(INFO, "MPI Paths have diverged: I am trying to read ", 
-                                name, " but some other process is trying to read ", 
-                                othername);
-    }
-
-    amrex::BoxArray ba = a_mf.boxArray();
-    amrex::DistributionMapping dm = a_mf.DistributionMap();
-    int nghost = a_mf.nGrow();
-    int ncomp = a_mf.nComp();
-    if (domain != amrex::Box::TheUnitBox())
-    {
-        for (auto &b : ba.boxList()) b = b.grow(2) & domain;
-        nghost = 0;
-    }
-    amrex::MultiFab mf(ba,dm,ncomp,nghost);
-    amrex::MultiFab::Copy(mf,a_mf,0,0,ncomp,0);
-    if (amrex::ParallelDescriptor::NProcs() == 1)
-    {
-        Util::ParallelMessage(INFO,"[",Compare_cnt,"] ","Storing ",name);
-        Util::ParallelMessage(INFO,"[",Compare_cnt,"] ",name, " - ncomp ",ncomp);
-        std::filesystem::create_directory("checks");
-        amrex::VisMF::Write(mf, "checks/" + name);
-    }
-    else
-    {
-        Util::ParallelMessage(INFO,"[",Compare_cnt,"] ","Comparing ",name);
-        {
-            amrex::MultiFab mftmp;
-            amrex::VisMF::Read(mftmp,"checks/" + name);
-            if (ba != mftmp.boxArray())
-            {
-                Util::ParallelMessage(file,func,line,"[",Compare_cnt,"] ","our boxarray ",ba);
-                Util::ParallelMessage(file,func,line,"[",Compare_cnt,"] ","saved boxarray ",mftmp.boxArray());
-                Util::Abort(file,func,line,"[",Compare_cnt,"] ","different box arrays !! !! !!");
-            }
-        }
-
-        amrex::MultiFab mforig(ba,mf.distributionMap,ncomp,nghost);
-        amrex::VisMF::Read(mforig,"checks/" + name);
-
-        amrex::MultiFab mfdiff(ba, mf.distributionMap, ncomp, nghost);
-        amrex::MultiFab::Copy(mfdiff, mf,0,0,ncomp,nghost);
-        amrex::MultiFab::Subtract(mfdiff,mforig,0,0,ncomp,nghost);
-        mfdiff.abs(0,ncomp,nghost);
-
-        for (int n = 0 ; n < ncomp; n++)
-        {
-            Set::Scalar max = mfdiff.max(n,nghost);
-            auto argmax = mfdiff.maxIndex(n,nghost);
-            if (max > tolerance)
-            {
-                amrex::ParallelDescriptor::Barrier();
-                Util::ParallelMessage(INFO,"ours (n=",n,")");
-                amrex::ParallelDescriptor::Barrier();
-                Util::Probe(INFO,mf,argmax[0],argmax[1],0,n,3);
-
-                amrex::ParallelDescriptor::Barrier();
-                Util::ParallelMessage(INFO,"original (n=",n,")");
-                amrex::ParallelDescriptor::Barrier();
-                Util::Probe(INFO,mforig,argmax[0],argmax[1],0,n,3);
-
-                amrex::ParallelDescriptor::Barrier();
-                Util::ParallelMessage(INFO,"diff (n=",n,")");
-                amrex::ParallelDescriptor::Barrier();
-                Util::Probe(INFO,mfdiff,argmax[0],argmax[1],0,n,3);
-
-                messagestream << "_diff";
-                amrex::VisMF::Write(mfdiff,messagestream.str());
-                Util::Warning(INFO,"[",Compare_cnt,"] ",name);
-                Util::Abort(INFO,"[",Compare_cnt,"] ",max," at ", argmax);
-            }
-        }
-    }
-    //if (Compare_cnt == 169) Util::ParallelAbort(INFO,"exiting");
-}
-
-
 std::string GetFileName()
 {
     if (filename == "")
